% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/baseball.R
\docType{data}
\name{baseball}
\alias{baseball}
\title{Birth dates of Major League Baseball players.}
\format{A data frame:
\describe{
  \item{name}{Player's name.}
  \item{birthdate}{Player's date of birth.}
  ...
}}
\source{
Data meticulously scraped from \url{http://www.baseball-reference.com/}.
}
\usage{
baseball
}
\description{
A dataset containing the birth dates of Major League Baseball players.
}
\keyword{datasets}

