% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/births.R
\docType{data}
\name{births}
\alias{births}
\title{Birth Statistics}
\format{A data frame:
\describe{
  \item{year}{Year.}
  \item{month}{Abbreviated month name.}
  \item{dow}{Day of week.}
  \item{sex}{Gender.}
  \item{count}{Number of births for specified year, month, day of week and sex.}
  ...
}}
\source{
Data gathered from http://www.cdc.gov/. Subsequently summarised.
}
\usage{
births
}
\description{
A dataset containing birth statistics for the US.
}
\keyword{datasets}

