% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deaths.R
\docType{data}
\name{deaths}
\alias{deaths}
\title{Death Statistics}
\format{A data frame:
\describe{
  \item{year}{Year.}
  \item{month}{Abbreviated month name.}
  \item{sex}{Gender.}
  \item{avgage}{Average age at death.}
  \item{medage}{Median age at death.}
  \item{maxage}{Maximum age at death.}
}}
\source{
Data gathered from http://www.cdc.gov/. Subsequently summarised.
}
\usage{
deaths
}
\description{
A dataset containing death statistics for the US.
}
\keyword{datasets}

