% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deathsage.R
\docType{data}
\name{deathsage}
\alias{deathsage}
\title{Death Counts by Year, Age and Gender}
\format{A data frame:
\describe{
  \item{year}{Year.}
  \item{age}{Age at death.}
  \item{sex}{Gender.}
  \item{count}{Number of deaths by age and sex.}
}}
\source{
Data gathered from http://www.cdc.gov/. Subsequently summarised.
}
\usage{
deathsage
}
\description{
A dataset containing death statistics for the US. Data are for the period 1994 to 2014.
}
\keyword{datasets}

