% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lifetable.R
\docType{data}
\name{lifetable}
\alias{lifetable}
\title{Life Table}
\format{A data frame:
\describe{
  \item{x}{Age in years.}
  \item{sex}{Gender.}
  \item{lx}{The number of people who survive to age x from an initial cohort of 100 000.}
  \item{dx}{The expected number of people from the initial cohort who die aged x on their last birthday.}
}}
\source{
Data gathered from http://www.cdc.gov/. Subsequently summarised.
}
\usage{
lifetable
}
\description{
Cohort life tables for each gender derived from CDC data.
}
\keyword{datasets}

