% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/population.R
\docType{data}
\name{population}
\alias{population}
\title{Population by Year}
\format{A data frame:
\describe{
  \item{year}{Year.}
  \item{age}{Age at death.}
  \item{sex}{Gender.}
  \item{count}{Number of deaths by age and sex.}
}}
\source{
Data gathered from http://www.multpl.com/united-states-population/.
}
\usage{
population
}
\description{
A dataset containing population statistics for the US. Data are for the period 1900 to 2016.
}
\keyword{datasets}

